src-python/requirements.txt
pyinstaller == 6.11.1
esptool == 4.8.1
pyserial == 3.5